users_ms/requirements.txt
django==3.1
djangorestframework==3.11.0
mysqlclient==2.0.0
django-cors-headers
